+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/EvaluationExplorer/EvaluationExplorerPage.tsx
@@ -6,9 +6,23 @@ import { ConfigurationBar } from './ConfigurationBar';
 import { EvaluationDataGrid } from './components';
 import { EvaluationExplorerPageProps, EvaluationRow } from './types';
 import { calculateRowDigest, deepCloneRow, createEmptyRow } from './utils';
-import SettingsIcon from '@mui/icons-material/Settings';
-import IconButton from '@mui/material/IconButton';
 
+// Constants
+const DEFAULT_DATASET_COLUMNS = ['columnA', 'columnB'];
+
+const createInitialRow = (): EvaluationRow => ({
+  id: "1",
+  dataset: {
+    columnA: "Value A1",
+    columnB: "Value B1",
+  },
+  output: {},
+  scores: {}
+});
+
+/**
+ * Main page component that wraps the inner explorer with header and layout.
+ */
 export const EvaluationExplorerPage = (props: EvaluationExplorerPageProps) => {
   const [configDrawerOpen, setConfigDrawerOpen] = useState(false);
 
@@ -23,7 +37,12 @@ export const EvaluationExplorerPage = (props: EvaluationExplorerPageProps) => {
       tabs={[
         {
           label: 'main',
-          content: <EvaluationExplorerPageInner {...props} configDrawerOpen={configDrawerOpen} onOpenConfigDrawer={handleOpenConfigDrawer} onCloseConfigDrawer={handleCloseConfigDrawer} />
+          content: <EvaluationExplorerPageInner 
+            {...props} 
+            configDrawerOpen={configDrawerOpen} 
+            onOpenConfigDrawer={handleOpenConfigDrawer} 
+            onCloseConfigDrawer={handleCloseConfigDrawer} 
+          />
         },
       ]}
       headerExtra={null}
@@ -31,38 +50,33 @@ export const EvaluationExplorerPage = (props: EvaluationExplorerPageProps) => {
   );
 };
 
-const DEFAULT_DATASET_COLUMNS = ['columnA', 'columnB'];
-
-const createInitialRow = (): EvaluationRow => ({
-  id: "1",
-  dataset: {
-    columnA: "Value A1",
-    columnB: "Value B1",
-  },
-  output: {},
-  scores: {}
-});
-
+/**
+ * Inner component that handles all the evaluation explorer logic and state.
+ */
 export const EvaluationExplorerPageInner: React.FC<EvaluationExplorerPageProps & {
   configDrawerOpen: boolean;
   onOpenConfigDrawer: () => void;
   onCloseConfigDrawer: () => void;
 }> = ({ entity, project, configDrawerOpen, onOpenConfigDrawer, onCloseConfigDrawer }) => {
+  // State management
   const [selectedDatasetId, setSelectedDatasetId] = useState<string>('dataset-1');
   const [isDatasetEdited, setIsDatasetEdited] = useState(false);
   const [selectedModelIds, setSelectedModelIds] = useState<string[]>([]);
   
+  // Data grid state
   const initialRow = createInitialRow();
   const [originalRows, setOriginalRows] = useState<GridRowsProp>([initialRow]);
   const [rows, setRows] = useState<GridRowsProp>([initialRow]);
   const [datasetColumns, setDatasetColumns] = useState<string[]>(DEFAULT_DATASET_COLUMNS);
   const [rowModesModel, setRowModesModel] = useState<GridRowModesModel>({});
 
+  // Check if dataset has been edited
   useEffect(() => {
     const hasBeenEdited = JSON.stringify(rows) !== JSON.stringify(originalRows);
     setIsDatasetEdited(hasBeenEdited);
   }, [rows, originalRows]);
 
+  // Update row digests whenever rows change
   useEffect(() => {
     setRows(prevRows => 
       prevRows.map(row => ({
@@ -75,6 +89,7 @@ export const EvaluationExplorerPageInner: React.FC<EvaluationExplorerPageProps &
     );
   }, [rows.length]);
 
+  // Row operations
   const handleDeleteRow = (id: GridRowId) => {
     setRows(prevRows => prevRows.filter((row) => row.id !== id));
   };
@@ -105,6 +120,7 @@ export const EvaluationExplorerPageInner: React.FC<EvaluationExplorerPageProps &
     });
   };
 
+  // Column operations
   const handleAddColumn = () => {
     setDatasetColumns(prevColumns => {
       const columnName = `column${prevColumns.length + 1}`;
@@ -134,6 +150,7 @@ export const EvaluationExplorerPageInner: React.FC<EvaluationExplorerPageProps &
     }));
   };
 
+  // Dataset operations
   const handleDatasetChange = (datasetId: string) => {
     if (datasetId === 'create-new') {
       setSelectedDatasetId('new-dataset');
@@ -143,6 +160,7 @@ export const EvaluationExplorerPageInner: React.FC<EvaluationExplorerPageProps &
       setDatasetColumns(DEFAULT_DATASET_COLUMNS);
     } else {
       setSelectedDatasetId(datasetId);
+      // TODO: Load dataset data from API
       const newRows = [{
         id: "1",
         dataset: {
@@ -161,18 +179,21 @@ export const EvaluationExplorerPageInner: React.FC<EvaluationExplorerPageProps &
     }
   };
 
+  // Model operations
   const handleModelsChange = (modelIds: string[]) => {
     setSelectedModelIds(modelIds);
     
     setRows(prevRows => prevRows.map(row => {
       const newOutput = { ...row.output };
       
+      // Add empty outputs for newly selected models
       modelIds.forEach(modelId => {
         if (!newOutput[modelId]) {
           newOutput[modelId] = null;
         }
       });
       
+      // Remove outputs for deselected models
       Object.keys(newOutput).forEach(modelId => {
         if (!modelIds.includes(modelId)) {
           delete newOutput[modelId];
@@ -186,6 +207,7 @@ export const EvaluationExplorerPageInner: React.FC<EvaluationExplorerPageProps &
   const handleRunModel = (rowId: string, modelId: string) => {
     console.log(`Running model ${modelId} for row ${rowId}`);
     
+    // TODO: Implement actual model invocation via API
     setTimeout(() => {
       setRows(prevRows => prevRows.map(row => {
         if (row.id === rowId) {
@@ -202,6 +224,7 @@ export const EvaluationExplorerPageInner: React.FC<EvaluationExplorerPageProps &
     }, 1000);
   };
 
+  // TODO: Replace with actual API data
   const availableModels = [
     { id: 'gpt-4', name: 'GPT-4', description: 'OpenAI GPT-4' },
     { id: 'claude-2', name: 'Claude 2', description: 'Anthropic Claude 2' },
@@ -242,6 +265,7 @@ export const EvaluationExplorerPageInner: React.FC<EvaluationExplorerPageProps &
   );
 };
 
+// Styled components
 const Container: React.FC<{ children: React.ReactNode }> = ({ children }) => (
   <div style={{
     width: '100%',
